feat: nixify installation process, reduce uncertainty
This commit replaces a lot of `add-apt-repository`, `curl` and `dpkg`
commands with a simple `nix profile add` or `flatpak install`, to ensure
increased stability, reproducibility and security. This way supply-chain
attacks are less viable and in general updates and maintenance will be
easier/trivial.

This wasn't done to packages which didn't require a third party
repo/install script (and ollama) as the main ubuntu repos are also
a very sensible way to install software.

Ollama wasn't changed, because that install script is non-trivial and
needs to install serveral drivers in order to work. I'll let the
maintainers of that package do their black magic on this one.